+++ torch_geometric_signed_directed/data/__init__.py
@@ -1,2 +1,2 @@
-from .signed_models import SSBM, polarized_ssbm
-from .directed_models import DSBM, meta_graph_generation, fix_network
+from .signed import *
+from .directed import *
