+++ torch_geometric_signed_directed/data/directed/__init__.py
@@ -1,0 +1,1 @@
+from .DSBM import DSBM
