+++ torch_geometric_signed_directed/utils/general/__init__.py
@@ -1,1 +1,2 @@
-from .data_utils import scipy_sparse_to_torch_sparse
+from .scipy_sparse_to_torch_sparse import scipy_sparse_to_torch_sparse
+from .extract_network import extract_network
